Move STARKVIND conveters to ikea device file
