+++ devices/ikea.js
@@ -8,6 +8,10 @@ const {repInterval} = require('../lib/constants');
 const extend = require('../lib/extend');
 const e = exposes.presets;
 const ea = exposes.access;
+const herdsman = require('zigbee-herdsman');
+const {
+    calibrateAndPrecisionRoundOptions, postfixWithEndpointName, getMetaValue,
+} = require('../lib/utils');
 
 const bulbOnEvent = async (type, data, device) => {
     /**
@@ -49,6 +53,152 @@ const tradfriExtend = {
     }),
 };
 
+const manufacturerOptions = {manufacturerCode: herdsman.Zcl.ManufacturerCode.IKEA_OF_SWEDEN};
+
+const ikea = {
+    fz: {
+        air_purifier: {
+            cluster: 'manuSpecificIkeaAirPurifier',
+            type: ['attributeReport', 'readResponse'],
+            options: [exposes.options.precision('pm25'), exposes.options.calibration('pm25')],
+            convert: (model, msg, publish, options, meta) => {
+                const state = {};
+
+                if (msg.data.hasOwnProperty('particulateMatter25Measurement')) {
+                    const pm25Property = postfixWithEndpointName('pm25', msg, model);
+                    let pm25 = parseFloat(msg.data['particulateMatter25Measurement']);
+
+                    // Air Quality Scale (ikea app):
+                    // 0-35=Good, 35-80=OK, 80+=Not Good
+                    let airQuality;
+                    const airQualityProperty = postfixWithEndpointName('air_quality', msg, model);
+                    if (pm25 <= 35) {
+                        airQuality = 'good';
+                    } else if (pm25 <= 80) {
+                        airQuality = 'ok';
+                    } else if (pm25 < 65535) {
+                        airQuality = 'not_good';
+                    } else {
+                        airQuality = 'unknown';
+                    }
+
+                    // calibrate and round pm25 unless invalid
+                    pm25 = (pm25 == 65535) ? -1 : calibrateAndPrecisionRoundOptions(pm25, options, 'pm25');
+
+                    state[pm25Property] = calibrateAndPrecisionRoundOptions(pm25, options, 'pm25');
+                    state[airQualityProperty] = airQuality;
+                }
+
+                if (msg.data.hasOwnProperty('filterRunTime')) {
+                // Filter needs to be replaced after 6 months
+                    state['replace_filter'] = (parseInt(msg.data['filterRunTime']) >= 259200);
+                }
+
+                if (msg.data.hasOwnProperty('controlPanelLight')) {
+                    state['led_enable'] = (msg.data['controlPanelLight'] == 0);
+                }
+
+                if (msg.data.hasOwnProperty('childLock')) {
+                    state['child_lock'] = (msg.data['childLock'] > 0 ? 'LOCK' : 'UNLOCK');
+                }
+
+                if (msg.data.hasOwnProperty('fanSpeed')) {
+                    let fanSpeed = msg.data['fanSpeed'];
+                    if (fanSpeed >= 10) {
+                        fanSpeed = (((fanSpeed - 5) * 2) / 10);
+                    } else {
+                        fanSpeed = 0;
+                    }
+
+                    state['fan_speed'] = fanSpeed;
+                }
+
+                if (msg.data.hasOwnProperty('fanMode')) {
+                    let fanMode = msg.data['fanMode'];
+                    if (fanMode >= 10) {
+                        fanMode = (((fanMode - 5) * 2) / 10).toString();
+                    } else if (fanMode == 1) {
+                        fanMode = 'auto';
+                    } else {
+                        fanMode = 'off';
+                    }
+
+                    state['fan_mode'] = fanMode;
+                    state['fan_state'] = (fanMode === 'off' ? 'OFF' : 'ON');
+                }
+
+                return state;
+            },
+        },
+    },
+    tz: {
+        air_purifier_fan_mode: {
+            key: ['fan_mode', 'fan_state'],
+            convertSet: async (entity, key, value, meta) => {
+                if (key == 'fan_state' && value.toLowerCase() == 'on') {
+                    value = getMetaValue(entity, meta.mapped, 'fanStateOn', 'allEqual', 'on');
+                }
+
+                let fanMode;
+                switch (value.toLowerCase()) {
+                case 'off':
+                    fanMode = 0;
+                    break;
+                case 'auto':
+                    fanMode = 1;
+                    break;
+                default:
+                    fanMode = parseInt(((parseInt(value) / 2.0) * 10) + 5);
+                }
+
+                await entity.write('manuSpecificIkeaAirPurifier', {'fanMode': fanMode}, manufacturerOptions.ikea);
+                return {state: {fan_mode: value.toLowerCase(), fan_state: value.toLowerCase() === 'off' ? 'OFF' : 'ON'}};
+            },
+            convertGet: async (entity, key, meta) => {
+                await entity.read('manuSpecificIkeaAirPurifier', ['fanMode']);
+            },
+        },
+        air_purifier_fan_speed: {
+            key: ['fan_speed'],
+            convertGet: async (entity, key, meta) => {
+                await entity.read('manuSpecificIkeaAirPurifier', ['fanSpeed']);
+            },
+        },
+        air_purifier_pm25: {
+            key: ['pm25', 'air_quality'],
+            convertGet: async (entity, key, meta) => {
+                await entity.read('manuSpecificIkeaAirPurifier', ['particulateMatter25Measurement']);
+            },
+        },
+        air_purifier_replace_filter: {
+            key: ['replace_filter'],
+            convertGet: async (entity, key, meta) => {
+                await entity.read('manuSpecificIkeaAirPurifier', ['filterRunTime']);
+            },
+        },
+        air_purifier_child_lock: {
+            key: ['child_lock'],
+            convertSet: async (entity, key, value, meta) => {
+                await entity.write('manuSpecificIkeaAirPurifier', {'childLock': ((value.toLowerCase() === 'lock') ? 1 : 0)},
+                    manufacturerOptions);
+                return {state: {child_lock: ((value.toLowerCase() === 'lock') ? 'LOCK' : 'UNLOCK')}};
+            },
+            convertGet: async (entity, key, meta) => {
+                await entity.read('manuSpecificIkeaAirPurifier', ['childLock']);
+            },
+        },
+        air_purifier_led_enable: {
+            key: ['led_enable'],
+            convertSet: async (entity, key, value, meta) => {
+                await entity.write('manuSpecificIkeaAirPurifier', {'controlPanelLight': ((value) ? 0 : 1)}, manufacturerOptions);
+                return {state: {led_enable: ((value) ? true : false)}};
+            },
+            convertGet: async (entity, key, meta) => {
+                await entity.read('manuSpecificIkeaAirPurifier', ['controlPanelLight']);
+            },
+        },
+    },
+};
 
 module.exports = [
     {
@@ -640,29 +790,28 @@ module.exports = [
                 .withDescription('Filter is older than 6 months and needs replacing'),
         ],
         meta: {fanStateOn: 'auto'},
-        fromZigbee: [fz.ikea_air_purifier],
+        fromZigbee: [ikea.fz.air_purifier],
         toZigbee: [
-            tz.ikea_air_purifier_fan_mode, tz.ikea_air_purifier_fan_speed,
-            tz.ikea_air_purifier_pm25, tz.ikea_air_purifier_child_lock, tz.ikea_air_purifier_led_enable,
-            tz.ikea_air_purifier_replace_filter,
+            ikea.tz.air_purifier_fan_mode, ikea.tz.air_purifier_fan_speed,
+            ikea.tz.air_purifier_pm25, ikea.tz.air_purifier_child_lock, ikea.tz.air_purifier_led_enable,
+            ikea.tz.air_purifier_replace_filter,
         ],
         configure: async (device, coordinatorEndpoint, logger) => {
-            const options = {manufacturerCode: 0x117c};
             const endpoint = device.getEndpoint(1);
 
             await reporting.bind(endpoint, coordinatorEndpoint, ['manuSpecificIkeaAirPurifier']);
             await endpoint.configureReporting('manuSpecificIkeaAirPurifier', [{attribute: 'particulateMatter25Measurement',
                 minimumReportInterval: repInterval.MINUTE, maximumReportInterval: repInterval.HOUR, reportableChange: 1}],
-            options);
+            manufacturerOptions);
             await endpoint.configureReporting('manuSpecificIkeaAirPurifier', [{attribute: 'filterRunTime',
                 minimumReportInterval: repInterval.HOUR, maximumReportInterval: repInterval.HOUR, reportableChange: 0}],
-            options);
+            manufacturerOptions);
             await endpoint.configureReporting('manuSpecificIkeaAirPurifier', [{attribute: 'fanMode',
                 minimumReportInterval: 0, maximumReportInterval: repInterval.HOUR, reportableChange: 1}],
-            options);
+            manufacturerOptions);
             await endpoint.configureReporting('manuSpecificIkeaAirPurifier', [{attribute: 'fanSpeed',
                 minimumReportInterval: 0, maximumReportInterval: repInterval.HOUR, reportableChange: 1}],
-            options);
+            manufacturerOptions);
 
             await endpoint.read('manuSpecificIkeaAirPurifier', ['controlPanelLight', 'childLock', 'filterRunTime']);
         },
